fix: refactor plugins
