+++ Sources/AmplitudeCore/Plugins/PluginHost.swift
@@ -1,0 +1,11 @@
+//
+//  PluginHost.swift
+//  AmplitudeCore
+//
+//  Created by Chris Leonavicius on 3/10/25.
+//
+
+protocol PluginHost {
+
+    func plugin(name: String) -> (UniversalPlugin)?
+}
